refactor(*): refactor name of env-getter module
